Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -3,7 +3,7 @@
 int main() {
     char A;
     scanf("%c",A);
-    int d=A;
+    int d=(int)A;
     printf("ASCII Value: %d",A);
     return 0;
 }
